🔄 update: Array constructor to use size_t for parameter

+++ module_7/ex02/Array.hpp
@@ -7,7 +7,7 @@ template <typename T>
 class Array {
 public:
     Array() : array(new T[0]), size(0) {}
-    Array(unsigned int n) : array(new T[n]), size(n) {}
+    Array(size_t n) : array(new T[n]), size(n) {}
     Array(const Array& other) : array(new T[other.size]), size(other.size) {
         for (size_t i = 0; i < size; i++) {
             array[i] = other.array[i];
